pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Youtube_Creator_Automation"
version = "0.1.0"
description = "Automated YouTube Shorts generation pipeline using GCP APIs."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.120.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "google-cloud-texttospeech>=2.16.0",
    "google-genai>=0.3.0",
    "openai-whisper>=20231117",
    "peewee>=3.17.0",
    "playwright>=1.40.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
